+++ src/main/java/acs/logic/db/DbUserService.java
@@ -17,6 +17,9 @@ import org.springframework.stereotype.Service;
 import acs.dal.UserDao;
 import acs.data.users.UserEntity;
 import acs.data.utils.UserIdEntity;
+import acs.data.utils.UserRole;
+import acs.logic.exceptions.EntityNotFoundException;
+import acs.logic.exceptions.ForbiddenActionException;
 import acs.logic.user.ExtentedUserService;
 import acs.logic.user.UserConvertor;
 import acs.rest.users.UserBoundary;
@@ -36,41 +39,39 @@ public class DbUserService implements ExtentedUserService {
 		this.projectName = projectName;
 	}
 
-	
-
 	@Autowired
 	public DbUserService(UserDao userDao, UserConvertor convertor, ValidEmail valid) {
 		this.userDao = userDao;
 		this.convertor = convertor;
 		this.valid = valid;
-		
+
 	}
 
 	@PostConstruct
 	public void init() {
 		// initialize object after injection
 		System.err.println("Project : " + this.projectName);
-		
+
 	}
 
 	@Override
 	@Transactional
 	public UserBoundary createUser(UserBoundary user) {
-		
+
 		if (!valid.isEmailVaild(user.getUserId().getEmail())) {
 			throw new RuntimeException("Email invalid!!");
 		}
-		
+
 		if (user.getAvatar() == null && user.getAvatar().trim().isEmpty()) {
 			throw new RuntimeException("Avatar invalid!!");
 		}
 		if (user.getUsername() == null) {
 			throw new RuntimeException("User Name invalid!!");
 		}
-		
+
 		user.setUserId(new UserIdBoundary(this.projectName, user.getUserId().getEmail()));
 		UserIdEntity userId = new UserIdEntity(user.getUserId().getDomain(), user.getUserId().getEmail());
-		
+
 		Optional<UserEntity> entityOptional = this.userDao.findById(userId);
 		if (entityOptional.isPresent()) {
 			throw new RuntimeException("the User alreay exsite in the data base!!");
@@ -88,11 +89,11 @@ public class DbUserService implements ExtentedUserService {
 	@Override
 	@Transactional
 	public UserBoundary login(String userDomain, String userEmail) {
-		
+
 		if (!(valid.isEmailVaild(userEmail))) {
 			throw new RuntimeException("Email invalid!!");
 		}
-		
+
 		UserIdEntity userId = new UserIdEntity(userDomain, userEmail);
 		Optional<UserEntity> entityOptional = this.userDao.findById(userId);
 		if (entityOptional.isPresent()) {
@@ -109,11 +110,11 @@ public class DbUserService implements ExtentedUserService {
 	@Override
 	@Transactional
 	public UserBoundary updateUser(String userDomain, String userEmail, UserBoundary update) {
-		
+
 		if (!valid.isEmailVaild(userEmail)) {
 			throw new RuntimeException("Email invalid!!");
 		}
-		
+
 		if (update.getAvatar() == null && update.getAvatar().trim().isEmpty()) {
 			throw new RuntimeException("Avatar invalid!!");
 		}
@@ -137,60 +138,50 @@ public class DbUserService implements ExtentedUserService {
 	@Override
 	@Transactional
 	public List<UserBoundary> getAllUsers(String adminDomain, String adminEmail) {
-		
 		if (!valid.isEmailVaild(adminEmail)) {
 			throw new RuntimeException("Email invalid");
 		}
-		
+
 		return StreamSupport.stream(this.userDao.findAll().spliterator(), false).map(this.convertor::fromEntity)
-						.collect(Collectors.toList());
+				.collect(Collectors.toList());
 
 	}
-	
+
 	@Override
 	@Transactional
 	public List<UserBoundary> getAllUsers(String adminDomain, String adminEmail, int page, int size) {
-		
+		UserEntity admin = this.userDao.findById(new UserIdEntity(adminDomain, adminEmail))
+				.orElseThrow(() -> new EntityNotFoundException(
+						"No User Exists With Domain: " + adminDomain + "Email: " + adminEmail));
+
+		if (!admin.getRole().equals(UserRole.ADMIN)) {
+			throw new ForbiddenActionException("Only Admin Has Permission To get all users");
+		}
+
 		if (!valid.isEmailVaild(adminEmail)) {
 			throw new RuntimeException("Email invalid");
 		}
 		return this.userDao.findAll(
-				// use pagination base on size & page and sort by ID in 
-				PageRequest.of(page, size, Direction.ASC, "userId"))
-				.getContent() // List<userBoudary>
+				// use pagination base on size & page and sort by ID in
+				PageRequest.of(page, size, Direction.ASC, "userId")).getContent() // List<userBoudary>
 				.stream() // Stream<userBoudary>
-				.map(this.convertor::fromEntity) //Stream <userBoudary>
+				.map(this.convertor::fromEntity) // Stream <userBoudary>
 				.collect(Collectors.toList());
 	}
 
 	@Override
 	@Transactional
 	public void deleteAllUsers(String adminDomain, String adminEmail) {
+		// Search the user that the action in Db
+		UserEntity admin = this.userDao.findById(new UserIdEntity(adminDomain, adminEmail))
+				.orElseThrow(() -> new EntityNotFoundException(
+						"No User Exists With Domain " + adminDomain + " Email " + adminEmail));
+
+		if (!admin.getRole().equals(UserRole.ADMIN)) {
+			throw new ForbiddenActionException("Only Admin Has Permission To delete Actions");
+		}
+
 		this.userDao.deleteAll();
-		/*
-		 * if (!valid.isEmailVaild(adminEmail)) { 
-		 * throw new RuntimeException("Email invalid!!"); 
-		 * } 
-		 * UserIdEntity userId = new UserIdEntity(adminDomain, adminEmail); 
-		 * Optional<UserEntity> entityOptional = this.userDao.findById(userId); 
-		 * 
-		 * if(entityOptional.isPresent()) {
-		 * 
-		 * 	 UserEntity entity = entityOptional.get();
-		 *  	 if (entity.getRole().equals(UserRole.ADMIN)) {
-		 * 					this.userDao.deleteAll(); 
-		 * } else { 
-		 *		 throw new RuntimeException("admin details invalid"); 
-		 * 	} 
-		 * }else{
-		 *		 throw new RuntimeException("No such admin details exsite"); 
-		 *}
-		 */
 
 	}
-
-
-
-	
-
 }
